UBER-838: button for signout in case the account is inactive
